ocr ajax

+++ application/daemon/ocr.php
@@ -6,10 +6,14 @@ foreach($files as $file){
     print($file);
     $tesseract = new TesseractOCR($file);
     $tesseract->setTempDir(realpath('temp'));
-    $result = $tesseract->recognize();
     $savefile = str_replace('pickup', 'ocr', $file);
     $savefile = str_replace('.jpg', '.txt', $savefile);
-    file_put_contents($savefile, $result);
+    if(file_exists($savefile)){
+
+    }else{
+        $result = $tesseract->recognize();
+        file_put_contents($savefile, $result);
+    }
 }
 
 ?>
